requirements.txt
# API core
fastapi[standard]==0.115.8
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Config / env
python-dotenv==1.0.1
pydantic-settings==2.7.1

# Google / Firebase
google-cloud-firestore==2.20.0
firebase-admin==6.6.0

# AI SDK
openai==1.61.1
openai>=1.10.0

# Monitoring
sentry-sdk[fastapi]==2.20.0

# Tests
pytest==8.3.4
pytest-mock>=3.10.0
httpx==0.28.1

# Lint
ruff==0.9.7
